requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
PyMuPDF==1.23.21
python-docx==1.1.0
spacy==3.7.2
sentence-transformers==2.5.1
pydantic==2.6.1
scikit-learn==1.4.0
numpy==1.26.4
streamlit>=1.30.0
requests>=2.31.0
pandas>=2.0.0
